refactor(ecospold): extract shared cut-off strategy into EcoSpold.Cutoff
The cut-off post-processing — reduce a multi-output dataset to a single
reference product so the engine sees a single-output process — was
duplicated verbatim between the EcoSpold1 and EcoSpold2 parsers. Move the
eight-function chain into one EcoSpold.Cutoff module both parsers import.

Tests that exercise these helpers now import them from EcoSpold.Cutoff
instead of the EcoSpold.Parser1 re-export.
